refactor: extract ThemeSwitch out of footer and add to header
