+++ components/theme-switch.tsx
@@ -1,0 +1,66 @@
+'use client'
+import { AnimatedBackground } from '@/components/ui/animated-background'
+import { MonitorIcon, MoonIcon, SunIcon } from 'lucide-react'
+import { useEffect, useState } from 'react'
+import { useTheme } from 'next-themes'
+
+const THEMES_OPTIONS = [
+  {
+    label: 'Light',
+    id: 'light',
+    icon: <SunIcon className="h-4 w-4" />,
+  },
+  {
+    label: 'Dark',
+    id: 'dark',
+    icon: <MoonIcon className="h-4 w-4" />,
+  },
+  {
+    label: 'System',
+    id: 'system',
+    icon: <MonitorIcon className="h-4 w-4" />,
+  },
+]
+
+export default function ThemeSwitch() {
+  const [mounted, setMounted] = useState(false)
+  const { theme, setTheme } = useTheme()
+
+  useEffect(() => {
+    setMounted(true)
+  }, [])
+
+  if (!mounted) {
+    return null
+  }
+
+  return (
+    <AnimatedBackground
+      className="pointer-events-none rounded-lg bg-zinc-100 dark:bg-zinc-800"
+      defaultValue={theme}
+      transition={{
+        type: 'spring',
+        bounce: 0,
+        duration: 0.2,
+      }}
+      enableHover={false}
+      onValueChangeAction={(id) => {
+        setTheme(id as string)
+      }}
+    >
+      {THEMES_OPTIONS.map((theme) => {
+        return (
+          <button
+            key={theme.id}
+            className="inline-flex h-7 w-7 items-center justify-center text-zinc-500 transition-colors duration-100 focus-visible:outline-2 data-[checked=true]:text-zinc-950 dark:text-zinc-400 dark:data-[checked=true]:text-zinc-50"
+            type="button"
+            aria-label={`Switch to ${theme.label} theme`}
+            data-id={theme.id}
+          >
+            {theme.icon}
+          </button>
+        )
+      })}
+    </AnimatedBackground>
+  )
+}
